-float to int update

+++ Product.cpp
@@ -3,12 +3,12 @@
 
 using namespace std;
 
-Product::Product(int id, const string& n, float p, int q)
+Product::Product(int id, const string& n, int p, int q)
 	:productID(id), name(n), price(p), quantity(q) {}
 
 int Product::getProductID() const { return productID; }
 string Product::getName() const { return name; }
-float Product::getPrice() const { return price; }
+int Product::getPrice() const { return price; }
 int Product::getQuantity() const { return quantity; }
 void Product::setQuantity(int q) { quantity = q; }
